Added I2C code for MPU-6050 module.
There is an issue with the SDA and SCL lines. After sending some data to
the module, the SDA line sometimes stays LOW rather than enter its idle
state (HIGH). This leads to the code getting stuck because the master
device (PIC24) wants to pull SDA LOW to start sending more data, but the
line is already LOW. I will find a solution to this issue.