Changes to I2C code and added PCB files.
I fixed the issue with the MPU6050 module. The module was in sleep mode,
and I needed to write to the PWR_MGMT_1 register to disable sleep mode.
Once I did this, the sensor no longer gave only 0x00 as output.

Also, I added some files for making a PCB using Altium.